+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Add `alias` keyWord location to Parsed Participant Definition
+- Add `KeyWord` type and a module
 
 ## 1.0.0 - 2020-11-05
 - Initial implementation